Refactor; move HTW into a separate package
