+++ scheduler/src/main/java/dev/gkblt/sdr/scheduler/components/HierarchicalTimerWheel.java
@@ -1,4 +1,4 @@
-package dev.gkblt.sdr.scheduler.services;
+package dev.gkblt.sdr.scheduler.components;
 
 import dev.gkblt.sdr.scheduler.model.Job;
