+++ scheduler/src/test/java/dev/gkblt/sdr/scheduler/components/HierarchicalTimerWheelTests.java
@@ -1,5 +1,6 @@
-package dev.gkblt.sdr.scheduler.services;
+package dev.gkblt.sdr.scheduler.components;
 
+import dev.gkblt.sdr.scheduler.components.HierarchicalTimerWheel;
 import dev.gkblt.sdr.scheduler.model.Job;
 import org.junit.jupiter.api.Test;
